app/requirements.txt
asgiref==3.4.1
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.8
djangorestframework==3.12.4
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
model-bakery==1.3.3
packaging==21.0
pluggy==1.0.0
psycopg2-binary==2.9.1
py==1.10.0
pyparsing==3.0.1
pytest==6.2.5
pytest-django==4.4.0
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
transliterate==1.10.2
uritemplate==4.1.1
urllib3==1.26.7
